utils/eks-cluster/scenario-definition/files/requirements.txt
pymongo
python-dotenv
PyYAML
